Finished and moved email files to new package
